Crate tidechain_client

Structs

LagoonExecutorDispatch

NativeElseWasmExecutor

A generic CodeExecutor implementation that uses a delegate to determine wasm code equivalence and dispatch to native code when possible, falling back on WasmExecutor when not.

TidechainExecutorDispatch

Enums

Client

A client instance of Tidechain.

Traits

AbstractClient

Trait that abstracts over all available client implementations.

ClientHandle

A handle to a Tidechain client instance.

ExecuteWithClient

Execute something with the client instance.

RuntimeApiCollection

A set of APIs that Tidechain-like runtimes must implement.

Type Definitions

FullBackend

FullClient